Rename downloader to resourcedownloader for clarity
