+++ pkg/workers/document/document_worker.go
@@ -21,9 +21,9 @@ import (
 	"github.com/gardener/docforge/pkg/registry/repositoryhost"
 	"github.com/gardener/docforge/pkg/workers/document/frontmatter"
 	"github.com/gardener/docforge/pkg/workers/document/markdown"
-	"github.com/gardener/docforge/pkg/workers/downloader"
 	"github.com/gardener/docforge/pkg/workers/linkresolver"
 	"github.com/gardener/docforge/pkg/workers/linkvalidator"
+	"github.com/gardener/docforge/pkg/workers/resourcedownloader"
 	"github.com/gardener/docforge/pkg/writers"
 	"github.com/yuin/goldmark/ast"
 	"k8s.io/klog/v2"
@@ -32,7 +32,7 @@ import (
 // Worker represents document worker
 type Worker struct {
 	linkresolver linkresolver.Interface
-	downloader   downloader.Interface
+	downloader   resourcedownloader.Interface
 	validator    linkvalidator.Interface
 
 	writer writers.Writer
@@ -52,7 +52,7 @@ type docContent struct {
 }
 
 // NewDocumentWorker creates Worker objects
-func NewDocumentWorker(resourcesRoot string, downloader downloader.Interface, validator linkvalidator.Interface, linkResolver linkresolver.Interface, rh registry.Interface, hugo hugo.Hugo, writer writers.Writer, skipLinkValidation bool) *Worker {
+func NewDocumentWorker(resourcesRoot string, downloader resourcedownloader.Interface, validator linkvalidator.Interface, linkResolver linkresolver.Interface, rh registry.Interface, hugo hugo.Hugo, writer writers.Writer, skipLinkValidation bool) *Worker {
 	return &Worker{
 		linkResolver,
 		downloader,
