+++ pkg/workers/document/document_worker_test.go
@@ -17,9 +17,9 @@ import (
 	"github.com/gardener/docforge/pkg/registry"
 	"github.com/gardener/docforge/pkg/registry/repositoryhost"
 	"github.com/gardener/docforge/pkg/workers/document"
-	"github.com/gardener/docforge/pkg/workers/downloader/downloaderfakes"
 	"github.com/gardener/docforge/pkg/workers/linkresolver/linkresolverfakes"
 	"github.com/gardener/docforge/pkg/workers/linkvalidator/linkvalidatorfakes"
+	"github.com/gardener/docforge/pkg/workers/resourcedownloader/downloaderfakes"
 	"github.com/gardener/docforge/pkg/writers/writersfakes"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
