+++ pkg/workers/resourcedownloader/resourcedownloader.go
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package downloader
+package resourcedownloader
 
 import (
 	"context"
@@ -17,8 +17,8 @@ import (
 	"k8s.io/klog/v2"
 )
 
-// DownloadWorker is the structure that processes downloads
-type DownloadWorker struct {
+// ResourceDownloadWorker is the structure that processes downloads
+type ResourceDownloadWorker struct {
 	registry registry.Interface
 	writer   writers.Writer
 	// lock for accessing the downloadedResources map
@@ -28,14 +28,14 @@ type DownloadWorker struct {
 }
 
 // NewDownloader creates new downloader
-func NewDownloader(registry registry.Interface, writer writers.Writer) (*DownloadWorker, error) {
+func NewDownloader(registry registry.Interface, writer writers.Writer) (*ResourceDownloadWorker, error) {
 	if registry == nil || reflect.ValueOf(registry).IsNil() {
 		return nil, errors.New("invalid argument: reader is nil")
 	}
 	if writer == nil || reflect.ValueOf(writer).IsNil() {
 		return nil, errors.New("invalid argument: writer is nil")
 	}
-	return &DownloadWorker{
+	return &ResourceDownloadWorker{
 		registry:            registry,
 		writer:              writer,
 		downloadedResources: make(map[string]struct{}),
@@ -43,7 +43,7 @@ func NewDownloader(registry registry.Interface, writer writers.Writer) (*Downloa
 }
 
 // Download downloads source as target
-func (d *DownloadWorker) Download(ctx context.Context, source string, target string, document string) error {
+func (d *ResourceDownloadWorker) Download(ctx context.Context, source string, target string, document string) error {
 	if !d.shouldDownload(source) {
 		return nil
 	}
@@ -60,7 +60,7 @@ func (d *DownloadWorker) Download(ctx context.Context, source string, target str
 }
 
 // shouldDownload checks whether a download task for the same Source is being processed
-func (d *DownloadWorker) shouldDownload(Source string) bool {
+func (d *ResourceDownloadWorker) shouldDownload(Source string) bool {
 	d.mux.Lock()
 	defer d.mux.Unlock()
 	if _, ok := d.downloadedResources[Source]; ok {
@@ -70,7 +70,7 @@ func (d *DownloadWorker) shouldDownload(Source string) bool {
 	return true
 }
 
-func (d *DownloadWorker) download(ctx context.Context, Source string, Target string) error {
+func (d *ResourceDownloadWorker) download(ctx context.Context, Source string, Target string) error {
 	reosurceURL, err := d.registry.ResourceURL(Source)
 	if err != nil {
 		return err
